+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@acala-network/chopsticks-testing",
+	"name": "@acala-network/chopsticks-utils",
 	"version": "0.9.2-2",
 	"author": "Acala Developers <hello@acala.network>",
 	"license": "Apache-2.0",
@@ -12,15 +12,11 @@
 		"build:types": "tsc -p tsconfig.json --emitDeclarationOnly --outDir dist/types"
 	},
 	"dependencies": {
-		"@acala-network/chopsticks-utils": "workspace:*"
-	},
-	"peerDependencies": {
-		"vitest": "^0.34.0"
+		"@acala-network/chopsticks": "workspace:*"
 	},
 	"devDependencies": {
 		"@swc/cli": "0.1.62",
 		"@swc/core": "^1.3.96",
-		"@types/chai": "^4.3.10",
 		"typescript": "^5.2.2"
 	},
 	"files": [
