+++ packages/utils/tsconfig.json
@@ -6,5 +6,5 @@
 	},
 	"include": ["src/**/*"],
 	"exclude": ["src/**/*.test.ts"],
-	"references": [{ "path": "../utils/tsconfig.json" }]
+	"references": [{ "path": "../chopsticks/tsconfig.json" }]
 }
